feat: add audio controls layout
